Fixes after Pull Request review

+++ src/test/tv/codely/scala_http_api/module/course/domain/CourseLessonsStub.scala
@@ -3,7 +3,7 @@ package tv.codely.scala_http_api.module.course.domain
 import tv.codely.scala_http_api.module.shared.domain.IntStub
 
 object CourseLessonsStub {
-  def apply(value: BigDecimal): CourseLessons = CourseLessons(value)
+  def apply(value: Int): CourseLessons = CourseLessons(value)
 
   def random: CourseLessons = CourseLessons(IntStub.randomBetween(1, 100))
 }
